Add AlternativeMistSlowdown option
