+++ worlds/okamihd/RegionsData/r20d.py
@@ -1,14 +1,9 @@
 from typing import TYPE_CHECKING
-
-from rule_builder.rules import True_, Has
-from ..CheckIds import container_check_id, brush_check_id, shop_check_id
+from ..CheckIds import container_check_id
 from ..Enums.BrushTechniques import BrushTechniques
-from ..Enums.LocationType import LocationType
 from ..Enums.OkamiEnemies import OkamiEnemies
 from ..Enums.RegionNames import RegionNames, MapIds
-from ..Enums.WarpType import WarpType
-from ..Rules import gen_thunder_chest_rule
-from ..Types import ExitData, LocData, EventData, WarpData
+from ..Types import ExitData, LocData, EventData
 
 if TYPE_CHECKING:
     from .. import OkamiWorld
